backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
beautifulsoup4==4.13.5
readability-lxml==0.8.4.1
lxml_html_clean==0.4.2
python-dotenv==1.1.1
curl_cffi[async]==0.13.0
docker==7.1.0
fake-useragent==2.2.0
openai==1.106.1
anthropic==0.66.0
google-genai==1.33.0
google-api-core==2.25.1